Refactor build scripts to use Gradle version catalog
The existing build.gradle file was deleted and replaced with build.gradle.kts. The switch to Kotlin-based scripts allows the usage of version catalogs - a Gradle feature for centralized management of dependency versions. A libs.versions.toml file was introduced to keep all version details in one place. This change enhances ease of dependencies management, eradicates redundancy, and simplifies upgrades.